[REFACTOR][RUNTIME][OPENCL] Fix missed texture.h include consumers
Four files outside of src/runtime/opencl/ referenced texture.h via its
old path at src/runtime/texture.h. The move audit (commit 2) only checked
for direct includees of opencl_common.h; these callers include texture.h
directly and were not caught.

Update the include path in all four files to src/runtime/opencl/texture.h:
- src/relax/transform/static_plan_block_memory.cc
- src/s_tir/backend/adreno/inject_texture_alloc.cc
- src/s_tir/backend/adreno/texture_flatten.cc
- src/target/opencl/codegen_opencl.cc
