+++ src/s_tir/backend/adreno/texture_flatten.cc
@@ -33,7 +33,7 @@
 #include <unordered_map>
 
 #include "../../../arith/ir_visitor_with_analyzer.h"
-#include "../../../runtime/texture.h"
+#include "../../../runtime/opencl/texture.h"
 #include "../../../runtime/thread_storage_scope.h"
 
 namespace tvm {
